etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "etl", from = "src" }]


[tool.poetry.dependencies]
python = ">=3.12,<3.13"
unstructured = {extras = ["all-docs"], version = "^0.15.8"}
boto3 = "^1.34.149"
langchain = "^0.2.11"
langchain-anthropic = "^0.1.20"
langchain-aws = "^0.1.12"
langchain-community = "^0.2.10"
langchain-core = "^0.2.24"
langchain-openai = "^0.1.17"
neo4j = "^5.22.0"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
unstructured-ingest = "^0.0.8"
unstructured-client = "^0.25.5"
pypdf2 = "^3.0.1"
openai = "^1.43.0"
pdf2image = "^1.17.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
